switched to 64 bit ints. sigh

+++ include/splatt.h
@@ -1,9 +1,11 @@
 #ifndef SPLATT_SPLATT_H
 #define SPLATT_SPLATT_H
 
+#include <stddef.h>
+
 typedef double splatt_val_t;
 #define SS_VAL "%f"
-typedef unsigned int splatt_idx_t;
-#define SS_IDX "%u"
+typedef size_t splatt_idx_t;
+#define SS_IDX "%lu"
 
 #endif
